+++ rag_quest/worlds/__init__.py
@@ -4,6 +4,7 @@ from .exporter import WorldExporter
 from .importer import WorldImporter
 from .modules import Module, ModuleManifestError, ModuleRegistry, load_modules
 from .templates import STARTER_WORLDS
+from .validate import ValidationResult, validate_manifest
 
 __all__ = [
     "WorldExporter",
@@ -13,4 +14,6 @@ __all__ = [
     "ModuleRegistry",
     "ModuleManifestError",
     "load_modules",
+    "validate_manifest",
+    "ValidationResult",
 ]